Provide entry state to all review views
Move the EntryStateProvider and collection watching from
EntryCommentsView up into the shared ReviewView base, so every review
sidebar panel can resolve entry structure — such as the section a
comment subject lives in — not just the comments list.

REDMINE-21261
