+++ entry_types/scrolled/package/src/editor/views/EntryCommentsView.js
@@ -1,7 +1,6 @@
-import React, {useEffect} from 'react';
+import React from 'react';
 import I18n from 'i18n-js';
 
-import {EntryStateProvider, useEntryStateDispatch, watchCollections} from 'pageflow-scrolled/entryState';
 import {ThreadList, useLocatedCommentThreads, review} from 'pageflow-scrolled/review';
 
 import {ReviewView} from './ReviewView';
@@ -24,7 +23,6 @@ export const EntryCommentsView = ReviewView.extend({
   props() {
     const {entry, editor} = this.options;
     return {
-      entry,
       selectedSubject: entry.get('selectedCommentsSubject') || null,
       // Undefined for elements without a slate cursor (e.g. images);
       // an array (possibly empty) for textBlocks where Selection.js
@@ -37,13 +35,8 @@ export const EntryCommentsView = ReviewView.extend({
     };
   },
 
-  renderContent({entry, ...props}) {
-    return (
-      <EntryStateProvider seed={entry.scrolledSeed}>
-        <WatchEntryCollections entry={entry} />
-        <CommentsList {...props} />
-      </EntryStateProvider>
-    );
+  renderContent(props) {
+    return <CommentsList {...props} />;
   },
 
   _onSelectedChange() {
@@ -69,14 +62,6 @@ export const EntryCommentsView = ReviewView.extend({
     }
   }
 });
-
-function WatchEntryCollections({entry}) {
-  const dispatch = useEntryStateDispatch();
-
-  useEffect(() => watchCollections(entry, {dispatch}), [entry, dispatch]);
-
-  return null;
-}
 
 function CommentsList({selectedSubject, transientThreadIds, highlightedThreadId, onThreadClick, editor}) {
   const {chapters} = useLocatedCommentThreads();
