+++ entry_types/scrolled/package/src/editor/views/ReviewView.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useEffect} from 'react';
 import ReactDOM from 'react-dom';
 import Marionette from 'backbone.marionette';
 
@@ -7,17 +7,24 @@ import {
   ReviewMessageHandler,
   ScrollHighlightedThreadIntoViewProvider
 } from 'pageflow-scrolled/review';
+import {
+  EntryStateProvider,
+  useEntryStateDispatch,
+  watchCollections
+} from 'pageflow-scrolled/entryState';
 
 import styles from './ReviewView.module.css';
 
 // Base Marionette view for comment-related sidebar panels. Provides the
 // shared wiring: a container div, a ReviewMessageHandler bridging the
-// session to the preview iframe, and a ReviewStateProvider seeded from
-// the current session state. Subclasses implement `renderContent(props)`
-// to return the React subtree. Props are produced by `props()` (default:
-// empty) and re-evaluated whenever the subclass calls `rerender()` —
-// useful for re-rendering on backbone change events without requiring
-// React subscription hooks inside the rendered tree.
+// session to the preview iframe, a ReviewStateProvider seeded from the
+// current session state, and an EntryStateProvider so the rendered tree
+// can resolve entry structure (e.g. the section a comment subject lives
+// in). Subclasses implement `renderContent(props)` to return the React
+// subtree. Props are produced by `props()` (default: empty) and
+// re-evaluated whenever the subclass calls `rerender()` — useful for
+// re-rendering on backbone change events without requiring React
+// subscription hooks inside the rendered tree.
 export const ReviewView = Marionette.ItemView.extend({
   template: () => `<div class="${styles.container}"></div>`,
 
@@ -42,12 +49,15 @@ export const ReviewView = Marionette.ItemView.extend({
   },
 
   rerender() {
-    const session = this.options.entry.reviewSession;
+    const {entry} = this.options;
     ReactDOM.render(
-      <ReviewStateProvider initialState={session.state}>
-        <ScrollHighlightedThreadIntoViewProvider>
-          {this.renderContent(this.props())}
-        </ScrollHighlightedThreadIntoViewProvider>
+      <ReviewStateProvider initialState={entry.reviewSession.state}>
+        <EntryStateProvider seed={entry.scrolledSeed}>
+          <WatchEntryCollections entry={entry} />
+          <ScrollHighlightedThreadIntoViewProvider>
+            {this.renderContent(this.props())}
+          </ScrollHighlightedThreadIntoViewProvider>
+        </EntryStateProvider>
       </ReviewStateProvider>,
       this._containerEl()
     );
@@ -57,3 +67,11 @@ export const ReviewView = Marionette.ItemView.extend({
     return this.$el.find(`.${styles.container}`)[0];
   }
 });
+
+function WatchEntryCollections({entry}) {
+  const dispatch = useEntryStateDispatch();
+
+  useEffect(() => watchCollections(entry, {dispatch}), [entry, dispatch]);
+
+  return null;
+}
